difftest2: pass spirv as words, not bytes

+++ tests/difftests/tests/lib/src/scaffold/shader/mod.rs
@@ -7,7 +7,7 @@ pub use wgsl_shader::WgslComputeShader;
 /// Trait for shaders that can provide SPIRV bytes.
 pub trait SpirvShader {
     /// Returns the SPIRV bytes and entry point name.
-    fn spirv_bytes(&self) -> anyhow::Result<(Vec<u8>, String)>;
+    fn spirv_bytes(&self) -> anyhow::Result<(Vec<u32>, String)>;
 }
 
 /// Trait for shaders that can create wgpu modules.
